requirements.txt
scipy
numpy>=1.16.1
h5py
gym==0.10.0
arrow
matplotlib
pytest-xdist
requests
clint
pillow
boto
future
pyzmq
enum34
gputil
joblib
cloudpickle
pyarrow==0.12.1
pyglet
six
flask
cherrypy
paste
simple-pid
requests
python-box
sarge

# Install from Github, editable
#-e git://github.com/deepdrive/gist.git#egg=python-gist
#-e git://github.com/deepdrive/youtube-upload#egg=youtube-upload

# Install from GitHub, non-editable
git+git://github.com/deepdrive/gist.git#egg=python-gist==0.6.1.dev42
git+git://github.com/deepdrive/youtube-upload#egg=youtube-upload

# The deepdrive bindings are installed via install.py according to the VERSION file in this repo